pyproject.toml
[project]
name = "genetics-tutor"
version = "0.1.0"
description = "RAG chatbot for genetics education using OER textbooks + Claude API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "anthropic>=0.40.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
genetics-tutor = "rag_chatbot.chat:main"
genetics-ingest = "rag_chatbot.ingest:main"

[tool.uv]
dev-dependencies = []
